initial cut utility class re  #43
